Unique Service now can Schedule an Event that will execute at a set Date and end at another set Date. This seems to work for short time intervalls but should be tested for longer ones.
A new class DateStringConverter was added that contains methods to convert String to Date and  Date to String. It is needed as we are working with intents and it is difficult to pass an Object (Date) with intent extras. This is a simpler solution.

The Main Activity now contains a test Event that can be scheduled manually. It is important to note that the current implementation only works with event starts in the future.